add beneficiary from mpbank + s2qbank
